Fixed pin contention
The motors were stop-starting with the new code included This was tracked down to the horn() function. It was discovered that the music object uses pin0 as its default output, which is also shared by the motor driver. As horn() is called every loop, even when the horn wasn't on the code still calls music.stop(), causing the pin contention.